requirements.txt
Django==3.0.4
djangorestframework==3.11.0
requests==2.23.0
python-decouple==3.3
psycopg2==2.8.4
drf-yasg==1.17.1